+++ tests/forms-schema/forms-schema.test.ts
@@ -2444,7 +2444,7 @@ test('should error if page element has child page element', () => {
   )
 
   expect(error?.details[0].message).toContain(
-    '"elements[0].elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
+    '"elements[0].elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointAddressV3, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
   )
 })
 
@@ -2525,7 +2525,7 @@ test('should error if isMultiPage is set to false', () => {
   )
 
   expect(error?.details[0].message).toContain(
-    '"elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
+    '"elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointAddressV3, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
   )
 })
 
@@ -2679,7 +2679,7 @@ test('should error if isMultiPage is false even if all root elements are pages',
   )
 
   expect(error?.details[0].message).toBe(
-    '"elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
+    '"elements[0].type" must be one of [abn, apiNSWLiquorLicence, arcGISWebMap, autocomplete, barcodeScanner, boolean, bsb, calculation, camera, captcha, checkboxes, civicaNameRecord, civicaStreetName, compliance, date, datetime, draw, email, files, form, freshdeskDependentField, geoscapeAddress, googleAddress, heading, html, image, infoPage, location, lookupButton, number, pointAddress, pointAddressV3, pointCadastralParcel, radio, repeatableSet, section, select, summary, telephone, text, textarea, time]',
   )
 })